+++ common/types.ts
@@ -1,3 +1,28 @@
+import { CSSObject } from "styled-components"
+
+export type ColorPreset =
+  | "neutral"
+  | "danger"
+  | "success"
+  | "warning"
+  | "unsaved"
+
+export type CustomColor = string
+
+export interface ParsedColorTheme {
+  background: any
+  font: any
+}
+
+export type fontSizeType = 10 | 12 | 14 | 16 | 18 | 20
+
+export interface BaseComponentProps {
+  /** CSS classname to be added */
+  className?: string
+  /** CSS styles to be adeed */
+  style?: CSSObject
+}
+
 /**
  * styled-components will by default apply valid HTML props onto the DOM.
  * Refer: https://www.styled-components.com/docs/basics#passed-props
@@ -10,3 +35,4 @@ export interface IStyled<PropType> {
   theme?: any
   [htmlProp: string]: any
 }
+
