+++ components/Label/__tests__/Label.test.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import Label from "../"
 import { render, cleanup } from "react-testing-library"
-import { ThemeProvider } from "../../styles"
+import { ThemeProvider } from "../../../common/styles"
 import "jest-styled-components"
 import "jest-dom/extend-expect"
