--- components/Modal/common/styles.ts
@@ -1,8 +1,0 @@
-import * as theme from "../../styles/variables"
-
-export const border = `1px solid ${theme.shades.gray85}`
-export const borderRadius = "0.4rem"
-export const padding = {
-  vertical: "2.1rem",
-  horizontal: "2.8rem",
-}
