+++ components/Modal/components/Footer.tsx
@@ -1,16 +1,16 @@
 import styled from "styled-components"
-import { border, padding } from "../common/styles"
-import { IStyled } from "../../common/types"
+import { IStyled } from "../../../common/types"
 
 interface FooterProps {
   showBorder: boolean
 }
 
 const Footer = styled.div<IStyled<FooterProps>>`
-  padding: ${`${padding.vertical} ${padding.horizontal}`};
+  padding: ${({ theme: { modalPadding } }) =>
+    `${modalPadding.vertical} ${modalPadding.horizontal}`};
   justify-self: flex-end;
-  border-top: ${({ customProps: { showBorder } }) =>
-    showBorder ? border : `none`};
+  border-top: ${({ customProps: { showBorder }, theme: { modalBorder } }) =>
+    showBorder ? modalBorder : `none`};
 `
 
 export default Footer
