+++ components/Modal/components/Main.tsx
@@ -1,8 +1,8 @@
 import styled from "styled-components"
-import { padding } from "../common/styles"
 
 const Main = styled.div`
-  padding: ${`${padding.vertical} ${padding.horizontal} 0rem`};
+  padding: ${({ theme: { modalPadding } }) =>
+    `${modalPadding.vertical} ${modalPadding.horizontal} 0rem`};
   overflow-y: auto;
   flex: 1 1 auto;
 `
