+++ components/index.tsx
@@ -6,4 +6,4 @@ export { default as Switch } from "./Switch"
 export { default as Icon } from "./Icon"
 export { default as Breadcrumb } from "./Breadcrumb"
 export { default as Label } from "./Label"
-export { default as ThemeProvider } from "./styles"
+export { default as ThemeProvider } from "../common/styles"
