+++ jest.config.js
@@ -4,7 +4,7 @@ module.exports = {
     "^.+\\.tsx?$": "ts-jest"
   },
   moduleNameMapper: {
-    "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/components/_utils/__mocks__/fileMock.ts",
+    "\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/common/_utils/__mocks__/fileMock.ts",
     "\\.(css|less)$": "identity-obj-proxy"
   },
   preset: "ts-jest",
